requirements.txt


fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9



torch>=2.0
torchvision


pillow
opencv-python-headless
numpy
scikit-image



reportlab



tqdm
matplotlib
albumentations